Backwards in-compatible changes:

Export connector tableInfo catalog elements are now named by
corresponding table name and not tableId. This is not externally
visible but *requires old catalogs to be recompiled for 1.1.*

ELT data sources can now be added a processor at runtime, not just at
initialization time. The RawProcessor will accept and store these but
will not produce a new advertisement. An operator adding and dropping
exported tables must restart all ELT clients to see the new
advertisements.  In-progress ELT data from a dropped table is
preserved and can be fully polled by an ELT client before or after the
client restarts.

The EE no longer uses tableIds (table.getRelativeIndex()) outside of
CatalogDelegates and the VoltDBEngine class. This has two user visible
consequences:

   1. Constraint failures now log by table name, not table id. This
should be more user friendly.

   2. ELT data sources no longer expose tableIds directly. They now
expose an opaque identifier (that is really a combination of catalog
version and table id). This id has changed time from int to long. ELT
clients must change to accept this. I did not rev the ELT client
protocol version. (Maybe I should?)

Internal changes:

The EE gains a CatalogDelegate interface. A catalog delegate manages
EE runtime objects derived from a catalog. At the moment, there is
only a table delegate. The interface, though, allows extracting EE
configuration out of VoltDBEngine.

The EE catalog now tracks added, updated and deleted objects after an
update. Call getDeletedPaths() to retrieve a list of object paths that
were deleted. Call wasAdded() or wasUpdated() on objects in the new
catalog to learn if they were updated or added by the last catalog
change.

The CatalogDiffEngine is now a real object, not just a set of static
methods. When instantiated, the engine performs a diff. The object
allows access to the text of the diff, whether the diff is a supported
runtime change, and if not, a list of text errors describing which
changes are invalid.

The allowed diff is now calculated during the diff calculation, not
after the diff is completed. The calculation is made using symbolic
CatalogTypes, not text string paths. The new white list supports
denying add-and-drop and denying object-modification at a field level.

There additional testcases for old catalog diff functionality.

Catalogs are now versioned internally. This not externally visible.
Catalog contexts now track catalog version. The version is incremented
each time an UpdateApplicationCatalog is executed. This is necessary
as the system must be able to distinguish add-drop-add sequences of
the same table.

FastSerializer gains two static methods: readString() and
writeString(), that perform VoltDB wire protocol compliant string
serialization into an existing bytebuffer.

VoltProjectBuilder gains an addLiteralSchema() method. You can add DDL
to a VoltProjectBuilder by passing it as a string. For example,
builder.addLiteralSchema("CREATE TABLE T(...);").

Thread.interrupt() is no longer called when shutting down sites during
regression suite executions. This cleans up some shutdown race
conditions that caused intermittant failures.

Several misues of std::map operator[] are fixed in VoltDBEngine.

+++ src/frontend/org/voltdb/exportclient/ExportClientBase.java
@@ -39,11 +39,11 @@ public abstract class ExportClientBase implements Runnable {
     protected HashMap<String, ExportConnection> m_elConnections;
 
     // First hash by table, second by partition
-    private final HashMap<Integer, HashMap<Integer, ExportDataSink>> m_sinks;
+    private final HashMap<Long, HashMap<Integer, ExportDataSink>> m_sinks;
 
     public ExportClientBase()
     {
-        m_sinks = new HashMap<Integer, HashMap<Integer, ExportDataSink>>();
+        m_sinks = new HashMap<Long, HashMap<Integer, ExportDataSink>>();
         m_elConnections = new HashMap<String, ExportConnection>();
         m_servers = null;
     }
@@ -74,7 +74,7 @@ public abstract class ExportClientBase implements Runnable {
                           ", part ID: " + source.partitionId());
             // Put the ELDataSink in our hashed collection if it doesn't exist
             ExportDataSink sink = null;
-            int table_id = source.tableId();
+            long table_id = source.tableId();
             int part_id = source.partitionId();
             HashMap<Integer, ExportDataSink> part_map =
                 m_sinks.get(table_id);
